rearrange/rename query functions
