+++ dev/postgresToCassandra.js
@@ -6,10 +6,8 @@ var constants = require('dotaconstants');
 var db = require('../store/db');
 var cassandra = require('../store/cassandra');
 var redis = require('../store/redis');
-var playerCache = require('../store/playerCache');
 var utility = require('../util/utility');
 var async = require('async');
-var updateCache = playerCache.updateCache;
 var serialize = utility.serialize;
 const cluster = require('cluster');
 var bucket_size = 100000000;
