+++ svc/web.js
@@ -2,37 +2,38 @@
  * Worker serving as main web application
  * Serves web/API requests
  **/
-var config = require('../config');
-var constants = require('dotaconstants');
-var utility = require('../util/utility');
-var redis = require('../store/redis');
-var status = require('../store/buildStatus');
-var db = require('../store/db');
-var cassandra = config.ENABLE_CASSANDRA_MATCH_STORE_READ ? require('../store/cassandra') : undefined;
-var queries = require('../store/queries');
-var matches = require('../routes/matches');
-var hyperopia = require('../routes/hyperopia');
-var players = require('../routes/players');
-var api = require('../routes/api');
-var donate = require('../routes/donate');
-var mmstats = require('../routes/mmstats');
-var request = require('request');
-var compression = require('compression');
-var session = require('cookie-session');
-var path = require('path');
-var moment = require('moment');
-var async = require('async');
-var fs = require('fs');
-var express = require('express');
-var app = express();
-var example_match = JSON.parse(fs.readFileSync('./matches/frontpage.json'));
-var passport = require('passport');
-var api_key = config.STEAM_API_KEY.split(",")[0];
-var SteamStrategy = require('passport-steam').Strategy;
-var host = config.ROOT_URL;
-var querystring = require('querystring');
-var util = require('util');
-var rc_public = config.RECAPTCHA_PUBLIC_KEY;
+const config = require('../config');
+const constants = require('dotaconstants');
+const utility = require('../util/utility');
+const redis = require('../store/redis');
+const status = require('../store/buildStatus');
+const db = require('../store/db');
+const cassandra = config.ENABLE_CASSANDRA_MATCH_STORE_READ ? require('../store/cassandra') : undefined;
+const queries = require('../store/queries');
+const searchPlayer = require('../store/search').searchPlayer;
+const matches = require('../routes/matches');
+const hyperopia = require('../routes/hyperopia');
+const players = require('../routes/players');
+const api = require('../routes/api');
+const donate = require('../routes/donate');
+const mmstats = require('../routes/mmstats');
+const request = require('request');
+const compression = require('compression');
+const session = require('cookie-session');
+const path = require('path');
+const moment = require('moment');
+const async = require('async');
+const fs = require('fs');
+const express = require('express');
+const app = express();
+const example_match = JSON.parse(fs.readFileSync('./matches/frontpage.json'));
+const passport = require('passport');
+const api_key = config.STEAM_API_KEY.split(",")[0];
+const SteamStrategy = require('passport-steam').Strategy;
+const host = config.ROOT_URL;
+const querystring = require('querystring');
+const util = require('util');
+const rc_public = config.RECAPTCHA_PUBLIC_KEY;
 //PASSPORT config
 passport.serializeUser(function (user, done)
 {
@@ -317,7 +318,7 @@ app.get('/benchmarks/:hero_id?', function (req, res, cb)
     }
     else
     {
-        queries.getBenchmarks(db, redis,
+        queries.getHeroBenchmarks(db, redis,
         {
             hero_id: req.params.hero_id
         }, function (err, result)
@@ -334,7 +335,7 @@ app.get('/search', function (req, res, cb)
 {
     if (req.query.q)
     {
-        queries.searchPlayer(db, req.query.q, function (err, result)
+        searchPlayer(db, req.query.q, function (err, result)
         {
             if (err)
             {
